requirements.txt
dash==2.18.2
dash-bootstrap-components==1.6.0
plotly==5.24.1
pandas==2.2.3
pyarrow==18.0.0
requests==2.32.3
python-dotenv==1.0.1
tqdm==4.67.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.26.1
numpy==2.1.3
gunicorn==23.0.0
